chore(refactor): #1223 feedback comments incorporated for model imports in same directory
