+++ web/src/app/core/services/index.service.ts
@@ -1,5 +1,5 @@
 export * from './authentication.service';
 export * from './project.service';
 export * from './repository.service';
-export * from './user.service';
 export * from './spinner.service';
+export * from './user.service';
