+++ web/src/app/shared/models/activity.model.ts
@@ -1,4 +1,4 @@
-import { RepositoryModel } from './index.model';
+import { RepositoryModel } from './repository.model';
 
 export class ActivityModel {
     type: string;
